Move login page component (#178)
